# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Trading Journal & Analytics API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /trades:
    get:
      operationId: getTrades
      tags: [trades]
      summary: Get all trades
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: startDate
          in: query
          schema:
            type: string
        - name: endDate
          in: query
          schema:
            type: string
        - name: assetClass
          in: query
          schema:
            type: string
        - name: direction
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
        - name: setupTag
          in: query
          schema:
            type: string
        - name: symbol
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of trades
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TradesResponse"
    post:
      operationId: createTrade
      tags: [trades]
      summary: Create a new trade manually
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTradeRequest"
      responses:
        "201":
          description: Trade created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Trade"
  /trades/{id}:
    get:
      operationId: getTrade
      tags: [trades]
      summary: Get trade by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Trade detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Trade"
    patch:
      operationId: updateTrade
      tags: [trades]
      summary: Update trade enrichment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTradeRequest"
      responses:
        "200":
          description: Trade updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Trade"
    delete:
      operationId: deleteTrade
      tags: [trades]
      summary: Delete a trade
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Trade deleted
...
  /playbooks:
    get:
      operationId: getPlaybooks
      tags: [playbooks]
      summary: Get all playbooks
      responses:
        "200":
          description: List of playbooks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Playbook"
    post:
      operationId: createPlaybook
      tags: [playbooks]
      summary: Create a playbook
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePlaybookRequest"
      responses:
        "201":
          description: Playbook created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Playbook"
  /playbooks/{id}:
    patch:
      operationId: updatePlaybook
      tags: [playbooks]
      summary: Update a playbook
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePlaybookRequest"
      responses:
        "200":
          description: Playbook updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Playbook"
    delete:
      operationId: deletePlaybook
      tags: [playbooks]
      summary: Delete a playbook
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Playbook deleted
...
components:
  schemas:
...
    Trade:
      type: object
      properties:
        id:
          type: integer
        symbol:
          type: string
        assetClass:
          type: string
          enum: [STOCK, CRYPTO, OPTION, FUTURES, FOREX]
        direction:
          type: string
          enum: [LONG, SHORT]
        quantity:
          type: number
        entryPrice:
          type: number
        exitPrice:
          type: number
        entryTime:
          type: string
        exitTime:
          type: string
        durationSeconds:
          type: number
        grossPnl:
          type: number
        netPnl:
          type: number
        commissions:
          type: number
        fees:
          type: number
        slippage:
          type: number
        status:
          type: string
          enum: [OPEN, CLOSED, PENDING]
        importSource:
          type: string
          enum: [API, CSV, MANUAL]
        setupTag:
          type: string
          nullable: true
        mistakeTag:
          type: string
          nullable: true
        emotionTag:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
        rating:
          type: integer
          nullable: true
        rMultiple:
          type: number
          nullable: true
        createdAt:
          type: string
      required:
        - id
        - symbol
        - assetClass
        - direction
        - quantity
        - entryPrice
        - exitPrice
        - entryTime
        - exitTime
        - durationSeconds
        - grossPnl
        - netPnl
        - commissions
        - fees
        - slippage
        - status
        - importSource
        - createdAt
    CreateTradeRequest:
      type: object
      properties:
        symbol:
          type: string
        assetClass:
          type: string
          enum: [STOCK, CRYPTO, OPTION, FUTURES, FOREX]
        direction:
          type: string
          enum: [LONG, SHORT]
        quantity:
          type: number
        entryPrice:
          type: number
        exitPrice:
          type: number
        entryTime:
          type: string
        exitTime:
          type: string
        commissions:
          type: number
          default: 0
        fees:
          type: number
          default: 0
        slippage:
          type: number
          default: 0
        stopLoss:
          type: number
          nullable: true
        setupTag:
          type: string
          nullable: true
        mistakeTag:
          type: string
          nullable: true
        emotionTag:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
        rating:
          type: integer
          nullable: true
      required:
        - symbol
        - assetClass
        - direction
        - quantity
        - entryPrice
        - exitPrice
        - entryTime
        - exitTime
    UpdateTradeRequest:
      type: object
      properties:
        setupTag:
          type: string
          nullable: true
        mistakeTag:
          type: string
          nullable: true
        emotionTag:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
        rating:
          type: integer
          nullable: true
...
    TradesResponse:
      type: object
      properties:
        trades:
          type: array
          items:
            $ref: "#/components/schemas/Trade"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required:
        - trades
        - total
        - page
        - limit
...
    Playbook:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
          nullable: true
        assetClass:
          type: string
          nullable: true
        timeWindow:
          type: string
          nullable: true
        rules:
          type: object
          nullable: true
        tradeCount:
          type: integer
        winRate:
          type: number
        netPnl:
          type: number
        avgRMultiple:
          type: number
        createdAt:
          type: string
      required:
        - id
        - name
        - tradeCount
        - winRate
        - netPnl
        - avgRMultiple
        - createdAt
    CreatePlaybookRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
          nullable: true
        assetClass:
          type: string
          nullable: true
        timeWindow:
          type: string
          nullable: true
        rules:
          type: object
          nullable: true
      required:
        - name
    UpdatePlaybookRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
          nullable: true
        assetClass:
          type: string
          nullable: true
        timeWindow:
          type: string
          nullable: true
        rules:
          type: object
          nullable: true
...