# lib/api-zod/src/generated/types/tradesResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Trading Journal & Analytics API
 * OpenAPI spec version: 0.1.0
 */
import type { Trade } from "./trade";

export interface TradesResponse {
  trades: Trade[];
  total: number;
  page: number;
  limit: number;
}